requirements.txt
hikari[speedups]==2.0.0.dev109
hikari-crescent>=0.2.4
hikari-miru==1.1.1
python-dotenv>=0.20.0